+++ tests/farming/endpoints/new-harvest-period.ts
@@ -6,6 +6,7 @@ import {
   assertApproxCurrentSlot,
   getCurrentSlot,
   errLogs,
+  sleep,
 } from "../../helpers";
 
 export function test() {
@@ -74,22 +75,25 @@ export function test() {
           10
         )
       );
-      expect(logs).to.contain("InvalidSlot");
+      expect(logs).to.contain("HarvestPeriodMustStartAtOrAfterCurrentSlot");
+    });
+
+    it("fails if period length is zero", async () => {
+      const periodLength = 0;
+
+      const logs = await errLogs(
+        farm.newHarvestPeriod(harvestMint, 0, periodLength, 10)
+      );
+      expect(logs).to.contain("HarvestPeriodMustBeAtLeastOneSlot");
     });
 
     it("fails with wrong signer pda", async () => {
       const fakePda = Keypair.generate().publicKey;
 
       const logs = await errLogs(
-        farm.newHarvestPeriod(
-          harvestMint,
-          0,
-          (await getCurrentSlot()) + 100,
-          defTps,
-          {
-            signerPda: fakePda,
-          }
-        )
+        farm.newHarvestPeriod(harvestMint, 0, 100, defTps, {
+          signerPda: fakePda,
+        })
       );
 
       expect(logs).to.contain("A seeds constraint was violated");
@@ -98,12 +102,7 @@ export function test() {
     it("interprets 0 as to start from current slot", async () => {
       const tps = defTps;
       const farmInfoBefore = await farm.fetch();
-      await farm.newHarvestPeriod(
-        harvestMint,
-        0,
-        (await getCurrentSlot()) + 100,
-        tps
-      );
+      await farm.newHarvestPeriod(harvestMint, 0, 100, tps);
       const farmInfoAfter = await farm.fetch();
 
       const harvestsBefore = farmInfoBefore.harvests as any[];
@@ -134,7 +133,7 @@ export function test() {
       await farm.newHarvestPeriod(
         harvestMint,
         currentSlot + 50,
-        currentSlot + 50 + period1Length - 1,
+        period1Length,
         tps
       );
       const vaultAfterPeriod1 = await farm.harvestVaultAccount(harvestMint);
@@ -146,7 +145,7 @@ export function test() {
       await farm.newHarvestPeriod(
         harvestMint,
         currentSlot + 200,
-        currentSlot + 200 + period2Length - 1,
+        period2Length,
         tps
       );
       const vaultAfterPeriod2 = await farm.harvestVaultAccount(harvestMint);
@@ -157,11 +156,64 @@ export function test() {
       await farm.newHarvestPeriod(
         harvestMint,
         currentSlot + 100,
-        currentSlot + 100 + period3Length - 1,
+        period3Length,
         tps
       );
       const vaultAfterPeriod3 = await farm.harvestVaultAccount(harvestMint);
       expect(Number(vaultAfterPeriod3.amount)).to.eq(tps * period3Length);
+    });
+
+    it("adds scheduled launch even if there's a running period", async () => {
+      const currentSlot = await getCurrentSlot();
+      const tps = defTps;
+
+      const vaultBefore = await farm.harvestVaultAccount(harvestMint);
+      expect(Number(vaultBefore.amount)).to.eq(0);
+
+      const period1Length = 100;
+      await farm.newHarvestPeriod(harvestMint, 0, period1Length, tps);
+      const vaultAfterPeriod1 = await farm.harvestVaultAccount(harvestMint);
+      expect(Number(vaultAfterPeriod1.amount)).to.eq(tps * period1Length);
+
+      // wait for a few slots in period 1
+      await sleep(1000);
+
+      // we can schedule a new period because the current one is already running
+      const period2Length = 200;
+      await farm.newHarvestPeriod(
+        harvestMint,
+        currentSlot + 300,
+        period2Length,
+        tps
+      );
+      const vaultAfterPeriod2 = await farm.harvestVaultAccount(harvestMint);
+      expect(Number(vaultAfterPeriod2.amount)).to.eq(
+        tps * period1Length + tps * period2Length
+      );
+
+      // and we can still change the scheduled launch
+      const period3Length = 50;
+      await farm.newHarvestPeriod(
+        harvestMint,
+        currentSlot + 200,
+        period3Length,
+        tps
+      );
+      const vaultAfterPeriod3 = await farm.harvestVaultAccount(harvestMint);
+      expect(Number(vaultAfterPeriod3.amount)).to.eq(
+        tps * period1Length + tps * period3Length
+      );
+
+      const logs = await errLogs(
+        farm.newHarvestPeriod(
+          harvestMint,
+          // somewhere halfway through the current active period
+          currentSlot + period1Length / 2,
+          10,
+          tps
+        )
+      );
+      expect(logs).to.contain("CannotOverwriteOpenHarvestPeriod");
     });
   });
 }