+++ programs/farming/CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-06-19
+
+### Added
+
+- New error variants related to adding a harvest.
+
+### Changed
+
+- When creating a new harvest period, the admin specifies the start and the
+  length, rather than the start and the end slot.
+- Admin can now schedule a new harvest period even if there's a harvest
+  period in progress.
+
+### Removed
+
+- Error variant `InvalidSlot`
+
 ## [1.0.0] - 2022-06-17
 
 ### Changed